LU-5577 ptlrpc: fix comparison between signed and unsigned
Change return type and size argiments of lustre_msg_hdr_size(),
lustre_msg_buf{len,count}() and req_capsule_*_size() to __u32.
Change type of req_format->rf_idx and req_format->rf_fields.nr
to size_t. Also return zero for incorrect message magic instead
of -EINVAL. This will be more robust because of few of them after
LASSERTF(0, "...") and will not be returned. In the rest places
it return zero size instead of huge number after implicit
unsigned conversion.

Reviewed-on: http://review.whamcloud.com/12475
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/lustre_req_layout.h
@@ -83,7 +83,8 @@ void req_capsule_set(struct req_capsule *pill, const struct req_format *fmt);
 void req_capsule_client_dump(struct req_capsule *pill);
 void req_capsule_server_dump(struct req_capsule *pill);
 void req_capsule_init_area(struct req_capsule *pill);
-int req_capsule_filled_sizes(struct req_capsule *pill, enum req_location loc);
+size_t req_capsule_filled_sizes(struct req_capsule *pill,
+				enum req_location loc);
 int  req_capsule_server_pack(struct req_capsule *pill);
 
 void *req_capsule_client_get(struct req_capsule *pill,
@@ -92,30 +93,30 @@ void *req_capsule_client_swab_get(struct req_capsule *pill,
 				  const struct req_msg_field *field,
 				  void *swabber);
 void *req_capsule_client_sized_get(struct req_capsule *pill,
-                                   const struct req_msg_field *field,
-                                   int len);
+				   const struct req_msg_field *field,
+				   __u32 len);
 void *req_capsule_server_get(struct req_capsule *pill,
                              const struct req_msg_field *field);
 void *req_capsule_server_sized_get(struct req_capsule *pill,
-                                   const struct req_msg_field *field,
-                                   int len);
+				   const struct req_msg_field *field,
+				   __u32 len);
 void *req_capsule_server_swab_get(struct req_capsule *pill,
                                   const struct req_msg_field *field,
                                   void *swabber);
 void *req_capsule_server_sized_swab_get(struct req_capsule *pill,
 					const struct req_msg_field *field,
-					int len, void *swabber);
+					__u32 len, void *swabber);
 const void *req_capsule_other_get(struct req_capsule *pill,
                                   const struct req_msg_field *field);
 
 void req_capsule_set_size(struct req_capsule *pill,
-                          const struct req_msg_field *field,
-                          enum req_location loc, int size);
-int req_capsule_get_size(const struct req_capsule *pill,
-                          const struct req_msg_field *field,
-                          enum req_location loc);
-int req_capsule_msg_size(struct req_capsule *pill, enum req_location loc);
-int req_capsule_fmt_size(__u32 magic, const struct req_format *fmt,
+			  const struct req_msg_field *field,
+			  enum req_location loc, __u32 size);
+__u32 req_capsule_get_size(const struct req_capsule *pill,
+			   const struct req_msg_field *field,
+			   enum req_location loc);
+__u32 req_capsule_msg_size(struct req_capsule *pill, enum req_location loc);
+__u32 req_capsule_fmt_size(__u32 magic, const struct req_format *fmt,
                          enum req_location loc);
 void req_capsule_extend(struct req_capsule *pill, const struct req_format *fmt);
 
@@ -126,12 +127,12 @@ int req_capsule_field_present(const struct req_capsule *pill,
                               const struct req_msg_field *field,
                               enum req_location loc);
 void req_capsule_shrink(struct req_capsule *pill,
-                        const struct req_msg_field *field,
-                        unsigned int newlen,
-                        enum req_location loc);
+			const struct req_msg_field *field,
+			__u32 newlen,
+			enum req_location loc);
 int req_capsule_server_grow(struct req_capsule *pill,
-                            const struct req_msg_field *field,
-                            unsigned int newlen);
+			    const struct req_msg_field *field,
+			    __u32 newlen);
 int  req_layout_init(void);
 void req_layout_fini(void);
